requirements.txt
streamlit
pandas
requests
fake_useragent
beautifulsoup4
weaviate-client
streamlit_float
google-cloud-storage
openai
lxml
python-dotenv
